Vector2 update for coordinates and speed

+++ Logic/IBallManager.cs
@@ -6,7 +6,7 @@ namespace Logic
 {
     public interface IBallManager
     {
-        void Create(int number, Double width, Double height, List<ILogicBallEvent> points);
+        void Create(int number, float width, float height, List<ILogicBallEvent> points);
         void Move();
         void MoveBall(IBall ball);
         void StopBalls();
